general commit

+++ app/src/main/java/com/example/playlistmaster/PlaylistSettingsActivity.kt
@@ -4,41 +4,36 @@ import CustomAdapter
 import android.annotation.SuppressLint
 import android.content.ClipData
 import android.content.ClipboardManager
+import android.content.DialogInterface
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.provider.MediaStore
 import android.provider.OpenableColumns
-import android.text.Layout
 import android.view.WindowManager
-import android.widget.FrameLayout
-import android.widget.ImageView
-import android.widget.LinearLayout
 import android.widget.TextView
 import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.AppCompatButton
-import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.core.net.toUri
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.google.gson.Gson
 import java.io.File
 import java.io.FileOutputStream
 import java.io.InputStream
-import java.util.ArrayList
-import kotlin.io.copyTo
+import java.util.Collections
 
 class PlaylistSettingsActivity : AppCompatActivity() {
 
     lateinit var playlist: File
     lateinit var unfinishedDir: File
+    lateinit var songOrderFile: File
     val audio = mutableListOf<File>()
     lateinit var customAdapter: CustomAdapter
-
     fun makeToast(msg: String) {
         Toast.makeText(this, msg, Toast.LENGTH_SHORT).show()
     }
@@ -50,7 +45,7 @@ class PlaylistSettingsActivity : AppCompatActivity() {
             val cursor = contentResolver.query(uri, null, null, null, null)
             cursor?.use {
                 if (it.moveToFirst()) {
-                    val columnIndex = it.getColumnIndex(OpenableColumns.DISPLAY_NAME) // Or MediaStore.Audio.Media.DATA
+                    val columnIndex = it.getColumnIndex(OpenableColumns.DISPLAY_NAME)
                     if (columnIndex != -1) {
                         filePath = it.getString(columnIndex)
                     }
@@ -60,6 +55,40 @@ class PlaylistSettingsActivity : AppCompatActivity() {
             filePath = uri.path
         }
         return filePath
+    }
+
+    fun setup() {
+        val playlistSerializable = intent.getSerializableExtra("playlist")
+        playlist = playlistSerializable as File
+        audio.clear()
+
+        songOrderFile = File(playlist, "songOrder.json")
+        JSON.getSongsOrdered(playlist).forEach { audio.add(File(playlist, it)) }
+
+        unfinishedDir = File(playlist, "unfinished")
+        if (!unfinishedDir.exists()) unfinishedDir.mkdir()
+    }
+
+
+    fun showConfirmationDialogue() {
+        val builder = AlertDialog.Builder(this)
+
+        builder.setTitle("Deletion confirmation")
+        builder.setMessage("Delete playlist ${playlist.name}? This action cannot be undone!")
+
+        builder.setPositiveButton("Delete") { dialog: DialogInterface, _: Int ->
+            playlist.deleteRecursively()
+            makeToast("Playlist ${playlist.name} deleted successfully!")
+            startActivity(Intent(this, MyPlaylistsActivity::class.java))
+            dialog.dismiss()
+        }
+
+        builder.setNegativeButton("Cancel") { dialog: DialogInterface, _: Int ->
+            dialog.dismiss()
+        }
+
+        val dialog: AlertDialog = builder.create()
+        dialog.show()
     }
 
     fun setupRecycler() {
@@ -77,8 +106,8 @@ class PlaylistSettingsActivity : AppCompatActivity() {
                 viewHolder: RecyclerView.ViewHolder
             ): Int {
                 val dragFlags =
-                    ItemTouchHelper.UP or ItemTouchHelper.DOWN // Allow vertical dragging
-                return makeMovementFlags(dragFlags, 0) // No swipe actions
+                    ItemTouchHelper.UP or ItemTouchHelper.DOWN
+                return makeMovementFlags(dragFlags, 0)
             }
 
             override fun onMove(
@@ -86,24 +115,16 @@ class PlaylistSettingsActivity : AppCompatActivity() {
                 viewHolder: RecyclerView.ViewHolder,
                 target: RecyclerView.ViewHolder
             ): Boolean {
-                // Update your data source (myDataSet) to reflect the new order
                 val fromPosition = viewHolder.adapterPosition
                 val toPosition = target.adapterPosition
-                // Perform the swap in your data list, e.g., Collections.swap(myDataSet, fromPosition, toPosition)
-                customAdapter.notifyItemMoved(
-                    fromPosition,
-                    toPosition
-                ) // Notify adapter of the move
+                Collections.swap(audio, fromPosition, toPosition)
+                customAdapter.notifyItemMoved(fromPosition, toPosition)
                 return true
             }
 
-            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                // Not used for drag-and-drop reordering
-            }
+            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {}
 
-            override fun isLongPressDragEnabled(): Boolean {
-                return true // Enable dragging by long press
-            }
+            override fun isLongPressDragEnabled(): Boolean { return true }
         }
 
         val itemTouchHelper = ItemTouchHelper(itemTouchHelperCallback)
@@ -111,53 +132,20 @@ class PlaylistSettingsActivity : AppCompatActivity() {
 
     }
 
-    fun setup() {
-        val playlistSerializable = intent.getSerializableExtra("playlist")
-        val audioSerializable = intent.getSerializableExtra("audio")
-
-        playlist = playlistSerializable as File
-        audio.clear()
-
-        if (audioSerializable is ArrayList<*>) {
-            audioSerializable.forEach {
-                audio.add(it as File)
-            }
-        }
-
-        unfinishedDir = File(playlist, "unfinished")
-        if (!unfinishedDir.exists()) unfinishedDir.mkdir()
-
-    }
-
     fun writeUnfinished(contentUri: Uri, fileName: String) {
-        var inputStream: InputStream? = null
-        var outputStream: FileOutputStream? = null
-        try {
-            // Open an InputStream from the content URI
-            inputStream = contentResolver.openInputStream(contentUri)
 
-            // Create a FileOutputStream for the destination file
-            outputStream = FileOutputStream(File(unfinishedDir, fileName))
+        val inputStream = contentResolver.openInputStream(contentUri)
+        val outputStream = FileOutputStream(File(unfinishedDir, fileName))
 
-            // Copy the data from the InputStream to the FileOutputStream
-            inputStream?.let { input ->
-                val buffer = ByteArray(4096) // Adjust buffer size as needed
-                var bytesRead: Int
-                while (input.read(buffer).also { bytesRead = it } != -1) {
-                    outputStream.write(buffer, 0, bytesRead)
-                }
+        inputStream?.let { input ->
+            val buffer = ByteArray(4096)
+            var bytesRead: Int
+            while (input.read(buffer).also { bytesRead = it } != -1) {
+                outputStream.write(buffer, 0, bytesRead)
             }
-        } catch (e: Exception) {
-            e.printStackTrace()
-            // Handle the exception (e.g., file not found, I/O error)
-        } finally {
-            // Close the streams
-            try {
-                inputStream?.close()
-                outputStream?.close()
-            } catch (e: Exception) {
-                e.printStackTrace()
-            }
+
+            inputStream.close()
+            outputStream.close()
         }
     }
     fun deleteUnfinished(fileName: String) {
@@ -166,12 +154,10 @@ class PlaylistSettingsActivity : AppCompatActivity() {
     }
 
     fun savePlaylist() {
+
         fun deleteRemovedFiles() {
             File(cacheDir, playlist.name).listFiles()?.forEach {
-                if (!audio.contains(it)) {
-                    it.delete()
-                    makeToast("deleted ${it.name}")
-                }
+                if (!audio.contains(it) && it.extension != "json") it.delete()
             }
         }
         fun createAddedFiles() {
@@ -185,6 +171,7 @@ class PlaylistSettingsActivity : AppCompatActivity() {
 
         deleteRemovedFiles()
         createAddedFiles()
+        songOrderFile.writeText(Gson().toJson(audio.map { it.name }))
 
         startActivity(Intent(this, MainActivity::class.java))
 
@@ -193,23 +180,12 @@ class PlaylistSettingsActivity : AppCompatActivity() {
     fun launchFileSelection() {
         var chooseFile = Intent(Intent.ACTION_OPEN_DOCUMENT)
         chooseFile.setType("audio/*")
-        chooseFile = Intent.createChooser(chooseFile, "Choose an audio file")
+        chooseFile = Intent.createChooser(chooseFile, "Choose audio to add to the playlist")
         startActivityForResult(chooseFile, 100)
     }
 
-
     @SuppressLint("SetTextI18n")
     override fun onCreate(savedInstanceState: Bundle?) {
-        Thread.setDefaultUncaughtExceptionHandler { thread, throwable ->
-            // Log the crash or show a toast
-            // Note: Toasts from background threads need to be posted to main thread
-            runOnUiThread {
-                Toast.makeText(this, "App crashed: ${throwable.cause}", Toast.LENGTH_LONG).show()
-                val clipboardManager = getSystemService(CLIPBOARD_SERVICE) as ClipboardManager
-                val clipData = ClipData.newPlainText("Copied Text", throwable.stackTraceToString())
-                clipboardManager.setPrimaryClip(clipData)
-            }
-        }
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContentView(R.layout.activity_playlist_settings)
@@ -223,10 +199,8 @@ class PlaylistSettingsActivity : AppCompatActivity() {
             insets
         }
 
-
         setup()
         setupRecycler()
-
 
         findViewById<TextView>(R.id.managePlaylistText).text = playlist.name
 
@@ -237,7 +211,11 @@ class PlaylistSettingsActivity : AppCompatActivity() {
         findViewById<AppCompatButton>(R.id.saveButtonM).setOnClickListener {
             savePlaylist()
         }
+        findViewById<AppCompatButton>(R.id.deleteButtonM).setOnClickListener {
+            showConfirmationDialogue()
+        }
     }
+
     override fun onActivityResult(
         requestCode: Int,
         resultCode: Int,
@@ -250,7 +228,7 @@ class PlaylistSettingsActivity : AppCompatActivity() {
 
             if (File(unfinishedDir, name).exists()) File(unfinishedDir, name).delete()
 
-            customAdapter.addAudio(File(unfinishedDir, name))
+            customAdapter.addAudio(File(unfinishedDir, name), uri)
         }
     }
